nothing

+++ detail.php
@@ -8,7 +8,26 @@ function numcheck($value){
         }
     }
 }
-
+function blockcheck($current_user){
+    $res = false;
+    $conn = OpenCon();
+    $query ="SELECT * FROM `admin_blocklist`";
+    //echo "Connected Successfully";
+    $result=$conn->query($query);
+    if ($result->num_rows > 0)
+    {
+   // OUTPUT DATA OF EACH ROW
+   while($row = $result->fetch_assoc())
+   {
+            if (strcmp($row['user_id'], $current_user->getName()) == 0) {
+                $res = true;
+                break;
+            } 
+   }
+    }
+    CloseCon($conn);
+    return $res;
+}
 include 'db_connection.php';
 require_once('session.php');
 session_start();
@@ -17,6 +36,7 @@ if (isset($_SESSION['current_user'])) {
 }else{
     $current_user=new User();
 }
+$block=blockcheck($current_user);
 // echo $current_user->getName();
 $conn = OpenCon();
 // echo $_SERVER['REQUEST_URI'];
@@ -64,20 +84,41 @@ if ($_SERVER["REQUEST_METHOD"] == "POST") {
         } else {
         //   echo "Error: " . $sql . "<br>" . $conn->error;
         }
-    CloseCon($conn);
+        CloseCon($conn);
     }
-if(isset($_POST['delete_id'])) {
-    $conn = OpenCon();
-    $delete_id = $_POST['delete_id'];
-    $delete = "DELETE FROM news_comment WHERE id = '$delete_id'";
-    if ($conn->query($delete) === TRUE) {
-        //   echo "Record deleted successfully";
-            $comment_count--;
-            update_comment($comment_count,$id);
-        } else {
-        //   echo "Error: " . $sql . "<br>" . $conn->error;
-    }
-    CloseCon($conn);
+    if (isset($_POST['admin_button'])) {
+        switch ($_POST['admin_button']) {
+            case "Delete":
+                if (isset($_POST['delete_id'])) {
+                    $conn = OpenCon();
+                    $delete_id = $_POST['delete_id'];
+                    $delete = "DELETE FROM news_comment WHERE id = '$delete_id'";
+                    if ($conn->query($delete) === TRUE) {
+                        echo "Record deleted successfully";
+                        $comment_count--;
+                        update_comment($comment_count, $id);
+                    } else {
+                        echo "Error: " . $sql . "<br>" . $conn->error;
+                    }
+                    CloseCon($conn);
+                }
+                break;
+            case "Block":
+                if (isset($_POST['block_name'])) {
+                    $conn = OpenCon();
+                    $block_name = $_POST['block_name'];
+                    // echo $block_name;
+                    $sql = "INSERT INTO admin_blocklist (user_id, blocked_by)
+        VALUES ('$block_name', '" . $current_user->getName() . "')";
+                    if ($conn->query($sql) === TRUE) {
+                        // echo "New record created successfully";
+                    } else {
+                        // echo "Error: " . $sql . "<br>" . $conn->error;
+                    }
+                    CloseCon($conn);
+                }
+                break;
+        }
     }
 }
 function update_comment($comment_count,$id){
@@ -153,7 +194,7 @@ CloseCon($conn);
                             <div class="card-body">
                                 <!-- Comment form-->
                                 <?php 
-                                if($current_user->getRole()!=0){
+                                if($current_user->getRole()!=0&&!$block){
                                     echo
                                         '<form class="mb-4" method="POST" action="">
                                     <textarea class="form-control" rows="3" placeholder="Join the discussion and leave a comment!" name="comment"></textarea>
@@ -221,12 +262,13 @@ CloseCon($conn);
                                         ' . $row['detail'] . '
                                         </div>
                                         <input type="hidden" name="delete_id" value="' . $row['id'] . '" />
-                                        <button type="submit" class="btn btn-danger pull-right mx-2" name="admin_button" value="delete">
-                                            Delete <i class="fa fa-arrow-circle-right"></i>
-                                        </button>
-                                        <button type="submit" class="btn btn-danger pull-right" name="admin_button" value="block">
-                                            Block <i class="fa fa-arrow-circle-right"></i>
-                                        </button>
+                                        <input type="hidden" name="block_name" value="' . $row['owner'] . '" />
+                                        <input type="submit" class="btn btn-danger pull-right mx-2" name="admin_button" value="Delete">
+                                            
+                                        </input>
+                                        <input type="submit" class="btn btn-danger pull-right" name="admin_button" value="Block">
+                                            
+                                        </input>
                                     </div>
                                     </form>';
                                             }else{
